Dios Aiudagit add -Agit add -A Dame inspiracion plox

+++ src/FGP_Layer/FoInterface.java
@@ -1,8 +1,8 @@
 package FGP_Layer;
 
-public interface FoInterface 
+public interface FoInterface
 {
-	void setIddleMinutes(int minutes);
+	void setIddleMinutes(long minutes);
 	void setId(String id);
 	void setDays(int days);
 	void setPeriod(int period);
